Add support for custom initializers, unordered_map-style emplace
Custom initializer can be passed as a second constructor argument, in
this case initializer type must be specified as a third template
argument for LruCache.
Initializer is a callable object accepting Key and returning Value.

emplace signature has been changed to be more similar to unordered_map.
It now returns a pair of a handle to the inserted element or to the
already-existing one if no insertion happened and a bool denoting
whether the insertion took place.